backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
easyocr
pytesseract
Pillow
pandas
sentence-transformers
faiss-cpu
numpy
rapidfuzz
torch